+++ example/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.owldemoapp",
+    package = "com.owldemo",
 )
 
 android_resource(
     name = "res",
-    package = "com.owldemoapp",
+    package = "com.owldemo",
     res = "src/main/res",
 )
