+++ example/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">OwlDemoApp</string>
+    <string name="app_name">OwlDemo</string>
 </resources>
