+++ example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "OwlDemoApp",
-  "displayName": "OwlDemoApp"
+  "name": "OwlDemo",
+  "displayName": "OwlDemo"
 }
