+++ example/ios/Podfile
@@ -3,7 +3,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '11.0'
 
-target 'OwlDemoApp' do
+target 'OwlDemo' do
   config = use_native_modules!
 
   use_react_native!(
@@ -12,7 +12,7 @@ target 'OwlDemoApp' do
     :hermes_enabled => false
   )
 
-  target 'OwlDemoAppTests' do
+  target 'OwlDemoTests' do
     inherit! :complete
     # Pods for testing
   end
