+++ example/owl.config.json
@@ -1,12 +1,12 @@
 {
   "ios": {
-    "workspace": "ios/OwlDemoApp.xcworkspace",
-    "scheme": "OwlDemoApp",
+    "workspace": "ios/OwlDemo.xcworkspace",
+    "scheme": "OwlDemo",
     "device": "iPhone 13 Pro",
     "quiet": true
   },
   "android": {
-    "packageName": "com.owldemoapp",
+    "packageName": "com.owldemo",
     "quiet": true
   },
   "debug": true
