+++ example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "owldemoapp",
+  "name": "owldemo",
   "version": "0.0.1",
   "private": true,
   "scripts": {
